% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoSim.R
\name{clamp}
\alias{clamp}
\title{Clamp \code{x} to be between \code{minVal} and \code{maxVal}}
\usage{
clamp(x, maxVal, minVal = 0)
}
\arguments{
\item{x}{value or vector of values to clamp}

\item{maxVal}{maximum value elements can take}

\item{minVal}{= 0 minimum value elements can take}
}
\value{
\code{x} if \code{x} is between \code{minVal} and \code{maxVal},
  otherwise \code{minVal} if \code{x} is lower or \code{maxVal} if \code{x}
  is higher
}
\description{
Clamp \code{x} to be between \code{minVal} and \code{maxVal}
}
