% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoSim.R
\name{connectAgents}
\alias{connectAgents}
\title{Return a connectivity matrix containing the connection strengths of the links
between each sender (cols) and receiver (rows) pair of \code{agents}}
\usage{
connectAgents(modelParams, agents, parents, oldTies)
}
\arguments{
\item{modelParams}{parameters for the model as a named list}

\item{agents}{data.frame of agents to be connected}

\item{parents}{list of vectors containing indices of previousGeneration
containing the parents to be used to produce each new agent.}

\item{oldTies}{the previous connectivity matrix used by parents}
}
\value{
n*n connectivity matrix where n is the number of rows in
  \code{agents}
}
\description{
This default function establishes reciprocal connections with a
  strength determined by the \code{\link{initalConnectionStrength}} function in
  \code{modelParams}
}
