% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoSim.R
\name{getAdvice}
\alias{getAdvice}
\title{Return the advice selected for each agent}
\usage{
getAdvice(modelParams, agents, world, ties)
}
\arguments{
\item{modelParams}{parameters for the model as a named list}

\item{agents}{data.frame of agents}

\item{world}{list of world properties (\code{state} obtained from
\code{\link{getWorldState}}, and \code{generation}, and \code{decision}
numbers.)}

\item{ties}{square matrix of connections between agents}
}
\value{
a version of the \code{agents} data frame including advice values.
}
\description{
Advisors are selected randomly from among the connections
  available to an agent. Custom replacements for this function may return
  more than one value for advice, provided \code{\link{getDecision}} is replaced with
  a function capable of handling the input
}
