% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoSim.R
\name{getDegree}
\alias{getDegree}
\title{Return the number of non-zero connections in row i of aMatrix}
\usage{
getDegree(i, aMatrix, threshold = 0)
}
\arguments{
\item{i}{row to inspect}

\item{aMatrix}{matrix containing at least \code{i} rows}

\item{threshold}{value whose magnitude connections in \code{i} must exceed to be
counted}
}
\value{
number of super-threshold values in row \code{i}
}
\description{
Return the number of non-zero connections in row i of aMatrix
}
