% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoSim.R
\name{initalConnectionStrength}
\alias{initalConnectionStrength}
\title{Return the initial strength of a connection between \code{sender} and \code{receiver}}
\usage{
initalConnectionStrength(receiver, sender)
}
\arguments{
\item{receiver}{data.frame row of the agent receiving information over the connection}

\item{sender}{data.frame row of the agent sending information over the connection}
}
\value{
the strength of the connection
}
\description{
Return the initial strength of a connection between \code{sender} and \code{receiver}
}
