% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoSim.R
\name{makeAgent}
\alias{makeAgent}
\title{Return an agent created from parents}
\usage{
makeAgent(modelParams, parents = NULL)
}
\arguments{
\item{modelParams}{parameters for the model as a named list}

\item{parents}{data.frame of parent agents who will have some say in
determining the properties of the agent}
}
\value{
\code{data.frame} row(s) representing the spawned agent
}
\description{
Return an agent created from parents
}
