% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoSim.R
\name{makeAgents}
\alias{makeAgents}
\title{Default function to define agents in the model}
\usage{
makeAgents(modelParams, previousGeneration = NULL, parents = NULL)
}
\arguments{
\item{modelParams}{parameters for the model as a named list}

\item{previousGeneration}{previous generation list containing data frame
$agents and connectivity matrix $ties. NULL for instantiating the first
generation}

\item{parents}{list of vectors containing indices of previousGeneration
containing the parents to be used to produce each new agent.}
}
\description{
This function is called each generation and returns a list with
  two components, a data frame of agents and a matrix of their connections.
  Replacements for this function should follow this format.
}
\section{returns list with two components}{
 \describe{
  \item{$agents}{\describe{
    \item{\code{id}}{global agent id}
    \item{\code{genId}}{agent id within this generation}
    \item{\code{generation}}{agent's generation}
    \item{\code{fitness}}{agent's fitness score}
    \item{...}{other fields defining the agent's specific properties}
  }}
  \item{$ties}{\code{n}*\code{n} matrix of connections between senders
   (column) and receivers (row) for each agent-agent pairing
  }}
}

