% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoSim.R
\name{selectParents}
\alias{selectParents}
\title{Select the parents for each slot in the next generation}
\usage{
selectParents(modelParams, agents, world, ties)
}
\arguments{
\item{modelParams}{parameters for the model as a named list}

\item{agents}{data.frame of agents}

\item{world}{list of world properties (\code{state} obtained from
\code{\link{getWorldState}}, and \code{generation}, and \code{decision}
numbers.)}

\item{ties}{square matrix of connections between agents}
}
\value{
a vector of agent ids for the parent of each to-be-spawned agent
}
\description{
Process the fitness of agents in a generation and select the
  ones which will reporoduce in the next generation. This implementation uses
  clones - only one parent exists for each agent, but other implementations
  can return a vector containing vectors of any length, allowing any number
  of parents for each agent. The output of this function is passed as a
  parameter to \code{\link{makeAgents}}
}
