% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoSim.R
\name{updateConnections}
\alias{updateConnections}
\title{Update the strength of connections between agents}
\usage{
updateConnections(modelParams, agents, ties, world)
}
\arguments{
\item{modelParams}{parameters for the model as a named list}

\item{agents}{data.frame of agents}

\item{ties}{square matrix of connections between agents}

\item{world}{list of world properties (\code{state} obtained from
\code{\link{getWorldState}}, and \code{generation}, and \code{decision}
numbers.)}
}
\value{
a square matrix similar to \code{ties} containing updated connection strengths
}
\description{
Agents' connection strengths can update dynamically. Customising
  this function will allow this to happen. This function is called as the
  last step in the decision process, after final decisions have been made and
  evaluated by \code{\link{getFitness}}.
}
