% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoSim.R
\docType{package}
\name{evoSim}
\alias{evoSim}
\alias{evoSim-package}
\title{evolutionary models of interacting agents}
\description{
egoSim allows for evolutionary models where the agents in any given
generation can interact with other agents via a connectivity matrix. This
allows, for example, agents to give one another advice on the solution to a
common problem.
}
\section{Functions}{
NA
}

\author{
Matt Jaquiery, \email{matt.jaquiery@psy.ox.ac.uk}
}
\keyword{algorithm,}
\keyword{evolution,}
\keyword{genetic}
\keyword{network}
\keyword{social}
