% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoSim.R
\name{getAdvisor}
\alias{getAdvisor}
\title{Return the advisor selected for each agent}
\usage{
getAdvisor(modelParams, agents, ties, world)
}
\arguments{
\item{modelParams}{parameters for the model as a named list}

\item{agents}{data.frame of agents}

\item{ties}{square matrix of connections between agents}

\item{world}{list of world properties (\code{state} obtained from
\code{\link{getWorldState}}, and \code{generation}, and \code{decision}
numbers.)}
}
\value{
a vector of agent ids identifying advisors which can be cound to the
  \code{agents} data frame.
}
\description{
Agents are selected randomly from among the connections
  available to an agent. Custom replacements for this function may return
  more than one advisor, provided \code{\link{getDecision}} is replaced with
  a function capable of handling the input
}
