% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoSim.R
\name{getDecision}
\alias{getDecision}
\title{Return the decisions made by agents}
\usage{
getDecision(modelParams, agents, world, ties, initial = F)
}
\arguments{
\item{modelParams}{parameters for the model as a named list}

\item{agents}{data.frame of agents}

\item{world}{list of world properties (\code{state} obtained from
\code{\link{getWorldState}}, and \code{generation}, and \code{decision}
numbers.)}

\item{ties}{square matrix of connections between agents}
}
\value{
an adjusted version of the \code{agents} data frame with decisions
  included.
}
\description{
This function uses the \code{agents$advisor} column to determine
  whether to produce an initial or final decision. Initial decisions are
  obtained by having an agent make a noisy assessment of the
  \code{world$state}, with a sensitivity based on its
  \code{agents$sensitivity}.

  Final decisions are made by taking a weighted average of the agent's
  initial decision and the initial decision of the selected advisor.
}
